Split up ProgressMessage into multiple smaller classes
