+++ calabash-adapter/pom.xml
@@ -11,7 +11,7 @@
   </parent>
 
   <artifactId>calabash-adapter</artifactId>
-  <version>2.2.0-SNAPSHOT</version>
+  <version>3.0.0-SNAPSHOT</version>
   <packaging>bundle</packaging>
 
   <name>DAISY Pipeline 2 :: Calabash adapter for the XProc API</name>
